chaos_mesh_experiments/requirements.txt
kubernetes>=28.1.0
PyYAML>=6.0.1 